chore(stage0): deduplicate imported fixture corpus
Result: {"status":"keep","redundant_fixture_groups":0,"imported_fixture_files":50,"wall_s":28.1}

+++ compiler/fixtures/programs/imported/typ/0003_annotated_value_float.ml
@@ -1,4 +1,4 @@
 fn main() {
-  let value: float = 2.0;
+  let value: f64 = 2.0;
   println(value)
 }
